+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 
 - Use Inter font in Dashboard
+- Hover over any place in the graph
 
 # 1.0.0 - Dec 12, 2025